fix: repair the three defects that made the core loss unusable
RkPinnLoss.forward raised RuntimeError on every call: the interpolant
einsum used "1" as a subscript label, which torch rejects. The result was
discarded into `_`, so the crash came from dead code but took the whole
library down with it.

The H1 initial-condition penalty called autograd.grad on u0, which is
supplied as sampled values and carries no autograd history. The target
derivative is now computed numerically on the x0 grid, with a
construction-time check that the grid is ordered so a non-monotonic x0
fails loudly instead of yielding a silently wrong derivative.

MLP built its layers in the global default dtype, so a model constructed
without a prior torch.set_default_dtype(torch.float64) met float64 inputs
and failed. Precision is now an explicit constructor argument.

Together these left the test suite red: 3 failed, and no configuration
of the library could evaluate a loss.

+++ src/pinn_rk/examples/train_heat_equation.py
@@ -53,7 +53,7 @@ def train_heat_equation(
         raise ValueError("Unknown RK method.")
 
     mesh = TimeMesh.uniform(T=T, N=N, device=device)
-    model = MLP(in_dim=2, width=128, depth=4, activation="tanh").to(device)
+    model = MLP(in_dim=2, width=128, depth=4, activation="tanh", dtype=torch.float64).to(device)
     L = Laplacian1D()
 
     x0, u0 = make_init_data(n0=128, device=device)
